Rename 'components/MixedBarAndLineChart.vue' to 'components/index/_shared/charts/MixedChart.vue'
