+++ components/index/_shared/charts/MixedChart.vue
